feat(MedicalCase): 医案聚合根CRUD重构
## 主要变更

### 统一保存端点
- 新增PUT /api/medicalcase/{id}/aggregate聚合根保存API
- 创建MedicalCaseAggregateInputDto统一Consultation+Prescription数据
- 事务保证诊断和处方原子性写入

### ISaveable到IDataProvider迁移
- 移除ISaveable接口依赖，使用IDataProvider模式
- ConsultationPanelViewModel实现IDataProvider<ConsultationInputDto>
- PrescriptionPanelViewModel实现IDataProvider<List<PrescriptionAggregateDto>>

### 工作区协调器优化
- MedicalCaseWorkspaceCoordinator统一收集子面板数据
- 移除独立的Consultation/Prescription保存API调用
- 保存API调用从2-3次减少到1次

### ConsultationModule禁用
- 从Shell层移除ConsultationModule注册
- 功能已迁移至MedicalCase模块的ConsultationPanelViewModel

OpenSpec: refactor-medicalcase-aggregate-crud

+++ src/Client/Desktop/Modules/LYBT.Desktop.MedicalCase/Services/PrescriptionCalculator.cs
@@ -113,7 +113,7 @@ public class PrescriptionCalculator
                 {
                     HerbId = item.HerbId,
                     HerbName = item.HerbName,
-                    Quantity = item.Dosage,
+                    Quantity = (int)item.Dosage,
                     Unit = item.Unit,
                     UnitPrice = unitPrice,
                     Subtotal = subtotal
